docs:v0.2.0リリース
Add new assets and terms documentation for v0.2.0 release

This commit includes the addition of several Excel files and a favicon for the new release under the `docs/assets/release/v0_2_0/jig/` directory. It also introduces a comprehensive `term.html` document containing a structured dictionary of terms related to the project.