chore:タスク
Update app link and add openApp task to Gulp workflow

Updated the application link in the navigation menu to point to the deployed URL instead of localhost. Added a new "openApp" Gulp task to launch the app in the default browser. Integrated "openApp" into the default Gulp workflow for improved development automation.